+++ API/Models/Gyroscope.cs
@@ -12,8 +12,8 @@ namespace ComEngineers.API.Models
         public int Id { get; set; }
         public DateTime TimeCode { get; set; }
         public required Session Session { get; set; }
-        public int XAxis { get; set; } // Pitch
-        public int YAxis { get; set; } // Yaw
-        public int ZAxis { get; set; } // Roll
+        public float XAxis { get; set; } // Pitch
+        public float YAxis { get; set; } // Yaw
+        public float ZAxis { get; set; } // Roll
     }
 }
